requirements.txt
-i https://pypi.org/simple
anyio==3.7.0 ; python_version >= '3.7'
attrs==22.2.0 ; python_version >= '3.6'
certifi==2023.5.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
fastapi==0.96.0
h11==0.14.0 ; python_version >= '3.7'
httptools==0.5.0
idna==3.4 ; python_version >= '3.5'
pydantic==1.10.8 ; python_version >= '3.7'
python-dotenv==1.0.0
pyyaml==6.0
requests==2.31.0 ; python_version >= '3.7'
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.27.0 ; python_version >= '3.7'
todoist-api-python==2.0.2
typing-extensions==4.6.3 ; python_version >= '3.7'
urllib3==2.0.2 ; python_version >= '3.7'
uvicorn[standard]==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
